refactoring: change package structure
