--- order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/OrderHistoryFacade.kt
@@ -1,19 +1,0 @@
-package org.heeheepresso.orderapi.orderHistory
-
-import org.heeheepresso.orderapi.orderHistory.dto.request.OrderHistoryCreateRequest
-import org.heeheepresso.orderapi.orderHistory.dto.response.OrderHistoryCreateResponse
-import org.springframework.stereotype.Component
-import org.springframework.transaction.annotation.Transactional
-
-@Component
-@Transactional(readOnly = true)
-class OrderHistoryFacade(
-    private val orderHistoryService: OrderHistoryService,
-    private val orderHistoryMapper: OrderHistoryMapper,
-) {
-
-    @Transactional
-    fun createOrderHistory(request: OrderHistoryCreateRequest): OrderHistoryCreateResponse =
-        orderHistoryMapper.convertToResponse(orderHistoryService.createOrderHistory(request))
-
-}
