--- order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/OrderHistoryRepositoryCustom.kt
@@ -1,6 +1,0 @@
-package org.heeheepresso.orderapi.orderHistory
-
-interface OrderHistoryRepositoryCustom {
-
-    fun findDetailById(id: Long): OrderHistory?
-}
