+++ order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/domain/model/OrderHistory.kt
@@ -1,11 +1,10 @@
-package org.heeheepresso.orderapi.orderHistory
+package org.heeheepresso.orderapi.orderHistory.domain.model
 
 import jakarta.persistence.*
 import org.heeheepresso.orderapi.common.BaseEntity
 import org.heeheepresso.orderapi.order.domain.model.OrderStatus
 import org.heeheepresso.orderapi.orderHistory.dto.request.OrderHistoryCreateRequest
-import org.heeheepresso.orderapi.orderHistory.menu.OrderMenuHistory
-import org.heeheepresso.orderapi.orderHistory.menu.dto.request.OrderMenuHistoryCreateRequest
+import org.heeheepresso.orderapi.orderHistory.dto.request.OrderMenuHistoryCreateRequest
 import java.math.BigDecimal
 
 @Entity
