+++ order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/domain/repository/OrderHistoryRepository.kt
@@ -1,5 +1,6 @@
-package org.heeheepresso.orderapi.orderHistory
+package org.heeheepresso.orderapi.orderHistory.domain.repository
 
+import org.heeheepresso.orderapi.orderHistory.domain.model.OrderHistory
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface OrderHistoryRepository : JpaRepository<OrderHistory, Long>, OrderHistoryRepositoryCustom
