+++ order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/domain/repository/OrderHistoryRepositoryCustom.kt
@@ -1,0 +1,8 @@
+package org.heeheepresso.orderapi.orderHistory.domain.repository
+
+import org.heeheepresso.orderapi.orderHistory.domain.model.OrderHistory
+
+interface OrderHistoryRepositoryCustom {
+
+    fun findDetailById(id: Long): OrderHistory?
+}
