+++ order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/domain/repository/OrderHistoryRepositoryImpl.kt
@@ -1,8 +1,7 @@
-package org.heeheepresso.orderapi.orderHistory.menu
+package org.heeheepresso.orderapi.orderHistory.domain.repository
 
 import com.querydsl.jpa.impl.JPAQueryFactory
-import org.heeheepresso.orderapi.orderHistory.OrderHistory
-import org.heeheepresso.orderapi.orderHistory.OrderHistoryRepositoryCustom
+import org.heeheepresso.orderapi.orderHistory.domain.model.OrderHistory
 import org.heeheepresso.orderapi.orderHistory.QOrderHistory.*
 import org.heeheepresso.orderapi.orderHistory.menu.QOrderMenuHistory.*
 import org.springframework.stereotype.Repository
