+++ order-api/src/main/kotlin/org/heeheepresso/orderapi/orderHistory/dto/request/OrderHistoryCreateRequest.kt
@@ -1,7 +1,6 @@
 package org.heeheepresso.orderapi.orderHistory.dto.request
 
 import org.heeheepresso.orderapi.order.domain.model.OrderStatus
-import org.heeheepresso.orderapi.orderHistory.menu.dto.request.OrderMenuHistoryCreateRequest
 import java.math.BigDecimal
 
 data class OrderHistoryCreateRequest(
